+++ lib/src/actions/action.js
@@ -1,6 +1,6 @@
 import { EventEmitter } from 'events';
 import { cloneDeep, has, noop } from 'lodash-es';
-import { TransformTypeClassMap, disabled, resolveAllKeys } from '../utils/index.js';
+import { Templating, Types, disabled } from '../utils/index.js';
 
 /**
  * @memberof module:Actions
@@ -34,13 +34,13 @@ class Action extends EventEmitter {
     if (this.obj.transforms) {
       // NOTE(jwetzell): turn transform JSON into class instances
       this.transforms = this.obj.transforms
-        .filter((transform) => has(TransformTypeClassMap, transform.type))
-        .map((transform) => new TransformTypeClassMap[transform.type](transform));
+        .filter((transform) => has(Types.TransformTypeClassMap, transform.type))
+        .map((transform) => new Types.TransformTypeClassMap[transform.type](transform));
     }
   }
 
   resolveTemplatedParams(data) {
-    return resolveAllKeys(this.params, data);
+    return Templating.resolveAllKeys(this.params, data);
   }
 
   /**
