+++ lib/src/actions/delay-action.js
@@ -1,5 +1,5 @@
 import { has } from 'lodash-es';
-import { ActionTypeClassMap, logger } from '../utils/index.js';
+import { Types, logger } from '../utils/index.js';
 import Action from './action.js';
 
 /**
@@ -12,8 +12,8 @@ class DelayAction extends Action {
 
     // NOTE(jwetzell): turn subAction JSON into class instances
     this.subActions = this.params.actions
-      .filter((action) => has(ActionTypeClassMap, action.type))
-      .map((action) => new ActionTypeClassMap[action.type](action));
+      .filter((action) => has(Types.ActionTypeClassMap, action.type))
+      .map((action) => new Types.ActionTypeClassMap[action.type](action));
   }
 
   doFunction(_msg, vars, protocols) {
