+++ lib/src/config.js
@@ -1,6 +1,6 @@
 import Ajv from 'ajv';
 import { cloneDeep, has } from 'lodash-es';
-import { MessageTypeClassMap, TriggerTypeClassMap } from './utils/index.js';
+import { Types } from './utils/index.js';
 
 class Config {
   constructor(configObj, schema) {
@@ -17,14 +17,14 @@ class Config {
   }
 
   loadTriggers() {
-    Object.keys(MessageTypeClassMap).forEach((messageType) => {
+    Object.keys(Types.MessageTypeClassMap).forEach((messageType) => {
       if (this.config[messageType]) {
         this[messageType] = this.config[messageType];
 
         // NOTE(jwetzell): turn trigger JSON into class instances
         this[messageType].triggers = this[messageType]?.triggers
-          ?.filter((trigger) => has(TriggerTypeClassMap, trigger.type))
-          .map((trigger) => new TriggerTypeClassMap[trigger.type](trigger));
+          ?.filter((trigger) => has(Types.TriggerTypeClassMap, trigger.type))
+          .map((trigger) => new Types.TriggerTypeClassMap[trigger.type](trigger));
       }
     });
   }
