+++ lib/src/protocols/cloud-protocol.js
@@ -1,7 +1,7 @@
 import { EventEmitter } from 'events';
 import { has } from 'lodash-es';
 import { io } from 'socket.io-client';
-import { MessageTypeClassMap, logger } from '../utils/index.js';
+import { Types, logger } from '../utils/index.js';
 
 class CloudProtocol extends EventEmitter {
   reload(params) {
@@ -40,8 +40,8 @@ class CloudProtocol extends EventEmitter {
       if (msgObj.messageType) {
         let message;
 
-        if (has(MessageTypeClassMap, msgObj.messageType)) {
-          message = MessageTypeClassMap[msgObj.messageType].fromJSON(msgObj);
+        if (has(Types.MessageTypeClassMap, msgObj.messageType)) {
+          message = Types.MessageTypeClassMap[msgObj.messageType].fromJSON(msgObj);
         } else {
           logger.error(`cloud: unhandled message type = ${msgObj.messageType}`);
         }
