+++ lib/src/transforms/template-transform.js
@@ -1,5 +1,5 @@
 import { set } from 'lodash-es';
-import { getTemplateResult, logger } from '../utils/index.js';
+import { Templating, logger } from '../utils/index.js';
 import Transform from './transform.js';
 
 /**
@@ -12,7 +12,7 @@ class TemplateTransform extends Transform {
     try {
       const resolvedParams = this.resolveTemplatedParams({ msg, vars });
 
-      let newValue = getTemplateResult(resolvedParams.template, { msg, vars });
+      let newValue = Templating.getTemplateResult(resolvedParams.template, { msg, vars });
       // NOTE(jwetzell): try to convert it to a number if it is one
       if (!Number.isNaN(parseFloat(newValue))) {
         newValue = parseFloat(newValue);
