+++ lib/src/transforms/transform.js
@@ -1,5 +1,5 @@
 import { noop } from 'lodash-es';
-import { disabled, resolveAllKeys } from '../utils/index.js';
+import { disabled, Templating } from '../utils/index.js';
 
 /**
  * @memberof module:Transforms
@@ -85,7 +85,7 @@ class Transform {
    * @returns {object} the params after running through templating engine
    */
   resolveTemplatedParams(data) {
-    return resolveAllKeys(this.params, data);
+    return Templating.resolveAllKeys(this.params, data);
   }
 
   /**
