+++ lib/src/triggers/trigger.js
@@ -1,5 +1,5 @@
 import { has } from 'lodash-es';
-import { ActionTypeClassMap, TriggerTypeClassMap, disabled, resolveAllKeys } from '../utils/index.js';
+import { Templating, Types, disabled } from '../utils/index.js';
 
 /**
  * @memberof module:Triggers
@@ -31,8 +31,8 @@ class Trigger {
       // NOTE(jwetzell): turn action JSON into class instances
       /** @type {module:Actions.Action[]} */
       this.actions = this.obj.actions
-        .filter((action) => has(ActionTypeClassMap, action.type))
-        .map((action) => new ActionTypeClassMap[action.type](action));
+        .filter((action) => has(Types.ActionTypeClassMap, action.type))
+        .map((action) => new Types.ActionTypeClassMap[action.type](action));
     }
   }
 
@@ -41,8 +41,8 @@ class Trigger {
       // NOTE(jwetzell): turn subTrigger JSON into class instances
       /** @type {module:Triggers.Trigger[]} */
       this.subTriggers = this.obj.subTriggers
-        .filter((subTrigger) => has(TriggerTypeClassMap, subTrigger.type))
-        .map((subTrigger) => new TriggerTypeClassMap[subTrigger.type](subTrigger));
+        .filter((subTrigger) => has(Types.TriggerTypeClassMap, subTrigger.type))
+        .map((subTrigger) => new Types.TriggerTypeClassMap[subTrigger.type](subTrigger));
     }
   }
 
@@ -101,7 +101,7 @@ class Trigger {
   }
 
   resolveTemplatedParams(data) {
-    return resolveAllKeys(this.params, data);
+    return Templating.resolveAllKeys(this.params, data);
   }
 
   /**
