+++ lib/src/utils/logging.js
@@ -1,0 +1,6 @@
+/* eslint-disable import/prefer-default-export */
+import pino from 'pino';
+
+const logger = pino();
+
+export default logger;
